[Chore #168849504] Restructured Errors in project
