+++ api/src/auth/products.js
@@ -1,7 +1,7 @@
 /* eslint-disable camelcase */
 import protocol from '../helpers/response';
 import database from '../db/pgConnect';
-import Errors from '../helpers/errors';
+import templateErrors from '../errors/templateLiterals';
 import Queries from '../queries/products';
 
 export default class Products {
@@ -9,7 +9,7 @@ export default class Products {
     const { name } = body;
     const findProductQuery = Queries.findProductByName();
     const findProduct = await database.queryOneORNone(findProductQuery, name);
-    if (findProduct) protocol.err404Res(res, Errors.dataFound('This item'));
+    if (findProduct) protocol.err404Res(res, templateErrors.dataFound('This item'));
     else next();
   }
 }
